added warps moved ContentGenerator to DataLayer
